Move graph_analysis module to graph.analysis
